perf: decouple koko types and move SFTP file transfer off host UI
Split koko shared types by domain, lift the file-transfer endpoint
   contract into connectors-core, and move the SFTP transfer store,
   checksum worker, and persistence into packages/koko. Remove the unused
   host TransferCenter overlay and the unused types/global.d.ts.
