+++ packages/connectors-core/package.json
@@ -6,8 +6,10 @@
   "exports": {
     ".": "./src/index.ts",
     "./types/component": "./src/types/component.ts",
+    "./types/fileTransfer": "./src/types/fileTransfer.ts",
     "./types/message": "./src/types/message.ts",
     "./types/session": "./src/types/session.ts",
+    "./fileTransferRegistry": "./src/fileTransferRegistry.ts",
     "./useConnectorEndpoint": "./src/useConnectorEndpoint.ts",
     "./useHostBridge": "./src/useHostBridge.ts",
     "./sessionContext": "./src/sessionContext.ts",
