+++ packages/koko/app/components/FileManagement/transfer-center/SftpTransferBatch.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
+import type { FileTransferConflictPolicy, FileTransferTask } from "@jumpserver/connectors-core";
 import type {
   SftpTransferBatchGroup,
   SftpTransferTargetGroup
 } from "#koko/composables/sftp/file-manager/transfer-center/useSftpTransferCenterSelectors";
 import type { SftpTransferGroupStatus } from "#koko/utils/sftpTransferSummary";
-import type { FileTransferConflictPolicy, FileTransferTask } from "~/shared/file-transfer/types";
 import SftpTransferTarget from "#koko/components/FileManagement/transfer-center/SftpTransferTarget.vue";
 import {
   batchHasFailedTasks,
