+++ packages/koko/app/tests/sftp/sftp.test.ts
@@ -1,4 +1,4 @@
-import type { FileTransferStatus, FileTransferTask } from "~/shared/file-transfer/types";
+import type { FileTransferStatus, FileTransferTask } from "@jumpserver/connectors-core";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import { ref } from "vue";
