+++ packages/koko/app/types/settings.ts
@@ -1,0 +1,10 @@
+export interface ITerminalSettings {
+  fontSize: number;
+  lineHeight: number;
+  fontFamily: string;
+  themeName: string;
+  ctrlCAsCtrlZ: string;
+  quickPaste: string;
+  backspaceAsCtrlH: string;
+  theme: string;
+}
