+++ packages/koko/app/utils/file-transfer/fileTransferChecksum.worker.ts
