--- types/global.d.ts
@@ -1,3 +1,0 @@
-declare interface ModuleOptions {
-  prefix: false | string
-}
